Move section to its own component
