+++ src/components/LandingSection/styles.ts
@@ -1,14 +1,6 @@
 import styled from "styled-components";
 import { motion } from "framer-motion";
 
-export const Section = styled.section`
-  height: 100vh;
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  flex-direction: column;
-`;
-
 export const PrimarySectionArea = styled.div`
   display: flex;
