+++ src/components/Section/styles.ts
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+export const Section = styled.section`
+  height: 100vh;
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+`;
